coin/gui/requirements.txt
PyQt5==5.15.10
qasync==0.28.0
aiohttp==3.9.1
websockets==12.0
mysql-connector-python==8.2.0
pandas==2.1.4
openpyxl==3.1.2
ccxt==4.5.2